Improve the way rendez-vous are removed from a group
If the user is no more in the group, we need to remove the rendez-vous he published in it, else he won't be able to access it anymore.
I do a check at 2 places :
- when building the rendez-vous group link
- when a rendez-vous group link was clicked

The second point can happen if the link was clicked from an email or an activity (edge case).

See #11